requirements.txt
Flask
python-dotenv
google-search-results
serpapi
pydantic
langchain-google-genai
langchain-core
langgraph
gunicorn
newspaper3kpip
lxml[html_clean]